+++ src/dependency_injector/types.py
@@ -1,4 +1,12 @@
 from typing import TypeVar, Generic, Any
+import warnings
+
+
+warnings.warn(
+    'Types module is deprecated since version 3.44.0. Use "providers" module instead: '
+    'providers.Provider[SomeClass]',
+    category=DeprecationWarning,
+)
 
 
 Injection = Any