+++ tests/unit/providers/test_types_py36.py
@@ -1,6 +1,6 @@
 import unittest
 
-from dependency_injector import providers, types
+from dependency_injector import providers
 
 
 class SomeClass:
@@ -10,6 +10,6 @@ class SomeClass:
 class TypesTest(unittest.TestCase):
 
     def test_provider(self):
-        provider: types.Provider[SomeClass] = providers.Factory(SomeClass)
+        provider: providers.Provider[SomeClass] = providers.Factory(SomeClass)
         some_object = provider()
         self.assertIsInstance(some_object, SomeClass)